requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-compute==1.14.0
